# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ypk btpit",
        "description": "Ypk btpit",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order/for-manager": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Получить заказы по роли (Админ видит все, Менеджер видит свои заказы)",
                "operationId": "fd2702d1d247f01bf485e1a775adf688",
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "executorId": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "customerId": {
                                                "type": "string"
                                            },
                                            "date": {
                                                "type": "string"
                                            },
                                            "statusName": {
                                                "type": "string"
                                            },
                                            "customersName": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "userComment": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "productDto": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/order/{order}": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Получить заказ по ID",
                "operationId": "663b00d3cebc9f4399e91e2025488c16",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "executorId": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "customerId": {
                                            "type": "string"
                                        },
                                        "date": {
                                            "type": "string"
                                        },
                                        "statusName": {
                                            "type": "string"
                                        },
                                        "customersName": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "userComment": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "productDto": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Удалить заказ по ID",
                "operationId": "35b2c18114e1dee54daeea6ee820e5bd",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {}
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order": {
            "post": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Создать заказ",
                "operationId": "a07ca1b7a7df6281e01b7b309254afb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "productId": {
                                        "type": "string"
                                    },
                                    "customersComment": {
                                        "type": "string"
                                    },
                                    "userComment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}